+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/ratelimit/BypassRateLimitRequestTokenSpec.groovy
@@ -2,8 +2,6 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web.ratelimit
 
-import com.yahoo.bard.webservice.web.RateLimitRequestToken
-
 import com.codahale.metrics.Meter
 
 import spock.lang.Specification