+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/web/ratelimit/CallbackRateLimitRequestTokenSpec.groovy
@@ -2,9 +2,6 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.web.ratelimit
 
-import com.yahoo.bard.webservice.web.RateLimitCleanupOnRequestComplete
-import com.yahoo.bard.webservice.web.RateLimitRequestToken
-
 import spock.lang.Specification
 
 class CallbackRateLimitRequestTokenSpec extends Specification {